Crate quick_protobuf

A library to read binary protobuf files

This reader is developed similarly to a pull reader

Re-exports

pub use crate::errors::Error;

pub use crate::errors::Result;

pub use crate::message::MessageInfo;

pub use crate::message::MessageRead;

pub use crate::message::MessageWrite;

pub use crate::reader::deserialize_from_slice;

pub use crate::reader::BytesReader;

pub use crate::writer::serialize_into_slice;

pub use crate::writer::BytesWriter;

pub use crate::writer::Writer;

pub use crate::writer::WriterBackend;

pub use crate::reader::Reader;

pub use crate::writer::serialize_into_vec;

Modules

errors

A module to handle all errors

message

A module to deserialize a Message as defined in a .proto file

reader

A module to manage protobuf deserialization

sizeofs

A module to compute the binary size of data once encoded

writer

A module to manage protobuf serialization