quick_protobuf::message

Trait MessageWrite

pub trait MessageWrite: Sized {
    // Provided methods
    fn write_message<W: WriterBackend>(&self, _: &mut Writer<W>) -> Result<()> { ... }
    fn get_size(&self) -> usize { ... }
    fn write_file<P: AsRef<Path>>(&self, p: P) -> Result<()> { ... }
}

A trait to handle deserialization based on parsed Fields

Provided Methods

fn write_message<W: WriterBackend>(&self, _: &mut Writer<W>) -> Result<()>

Writes Self into W writer

fn get_size(&self) -> usize

Computes necessary binary size of self once serialized in protobuf

fn write_file<P: AsRef<Path>>(&self, p: P) -> Result<()>

Writes self into a file

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.