quick_protobuf

Module reader

A module to manage protobuf deserialization

There are actually two main readers

a BytesReader which parses data from a &[u8]

a Reader which is a wrapper on BytesReader which has its own buffer. It provides convenient functions to the user suche as from_file

It is advised, for convenience to directly work with a Reader.

Structs

BytesReader

A struct to read protocol binary files

Reader

A struct to read protobuf data

Functions

deserialize_from_slice

Deserialize a MessageRead from a &u8`