Module quick_protobuf::sizeofs

A module to compute the binary size of data once encoded

This module is used primilarly when implementing the MessageWrite::get_size

Functions

sizeof_bool

Computes the binary size of the varint encoded bool (always = 1)

sizeof_enum

Computes the binary size of the varint encoded enum

sizeof_int32

Computes the binary size of the varint encoded i32

sizeof_int64

Computes the binary size of the varint encoded i64

sizeof_len

Computes the binary size of a variable length chunk of data (wire type 2)

sizeof_sint32

Computes the binary size of the varint encoded sint32

sizeof_sint64

Computes the binary size of the varint encoded sint64

sizeof_uint32

Computes the binary size of the varint encoded uint32

sizeof_uint64

Computes the binary size of the varint encoded uint64

sizeof_varint

Computes the binary size of the varint encoded u64