Trait quick_protobuf::message::MessageWrite

pub trait MessageWrite: Sized {
    // Provided methods
    fn write_message<W: WriterBackend>(&self, _: &mut Writer<W>) -> Result<()> { ... }
    fn get_size(&self) -> usize { ... }
    fn write_file<P: AsRef<Path>>(&self, p: P) -> Result<()> { ... }
}

A trait to handle deserialization based on parsed Fields

Provided Methods

fn write_message<W: WriterBackend>(&self, _: &mut Writer<W>) -> Result<()>

Writes Self into W writer

fn get_size(&self) -> usize

Computes necessary binary size of self once serialized in protobuf

fn write_file<P: AsRef<Path>>(&self, p: P) -> Result<()>

Writes self into a file

Object Safety

This trait is not object safe.